Refactor App.jsx
- Extract header, modal, and error overlay into components
- Centralise constants (timeouts, defaults, reticle, modal size)
- Switch header to React handlers; remove DOM listener wiring
- Make controls idempotent; guard keyboard listeners
- Move error overlay styles to CSS
