+++ gui/src/styles/components/App.css
@@ -53,10 +53,8 @@
 }
 
 #aladin-lite-div {
-  width: 100% !important;
-  height: 100% !important;
-  min-height: 400px !important;
-  position: relative !important;
+  min-height: 300px;
+  width: 100%;
 }
 
 #aladin-lite-div canvas {
@@ -195,4 +193,59 @@
   border-radius: 50%;
   animation: spin 1s linear infinite;
   margin-bottom: 1rem;
+}
+
+.error-overlay {
+	position: absolute;
+	top: 45%;
+	left: calc(50% + 7.5rem);
+	transform: translate(-50%, -50%);
+	background-color: #2D3748;
+	border: 2px solid #FC8181;
+	border-radius: 8px;
+	padding: 20px;
+	max-width: 400px;
+	text-align: center;
+	z-index: 1000;
+	box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
+}
+
+.error-overlay .error-title {
+	color: #FC8181;
+	font-size: 18px;
+	font-weight: 600;
+	margin-bottom: 12px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	gap: 8px;
+}
+
+.error-overlay .error-message {
+	color: #E2E8F0;
+	font-size: 14px;
+	margin-bottom: 12px;
+	line-height: 1.4;
+}
+
+.error-overlay .error-hint {
+	color: #A0AEC0;
+	font-size: 12px;
+	margin-bottom: 16px;
+}
+
+.error-overlay .retry-btn {
+	background-color: #4FD1C5;
+	color: #1A202C;
+	border: none;
+	border-radius: 4px;
+	padding: 8px 16px;
+	font-size: 14px;
+	font-weight: 600;
+	cursor: pointer;
+	transition: background-color 0.2s ease;
+}
+
+.error-overlay .retry-btn:hover {
+	background-color: #38B2AC;
 }
