+++ gui/src/utils/constants.js
@@ -1,0 +1,16 @@
+export const MAP_CONTROLS = {
+    'map-stellar-velocity': { label: 'Stellar velocity', icon: '🌀' },
+    'map-velocity-dispersion': { label: 'Velocity dispersion', icon: '📊' },
+    'map-stellar-velocity-error': { label: 'Stellar velocity Error', icon: '⚠️' },
+    'map-velocity-dispersion-error': { label: 'Velocity dispersion Error', icon: '📈' },
+    'map-h3': { label: 'h3', icon: 'H₃' },
+    'map-h4': { label: 'h4', icon: 'H₄' },
+    'map-age-weighted': { label: 'Age (lum. weighted)', icon: '⏳' },
+    'map-age-mass-weighted': { label: 'Age (Mass Weighted)', icon: '⚖️' },
+    'map-metallicity': { label: 'Metallicity', icon: '⚛️' }
+  };
+
+export const MODAL_DIMENSIONS = { widthPx: 928, heightPx: 500 };
+export const TIMEOUTS = { restoreGalleryMs: 100, hideCoordsMs: 1000, objectResolutionMs: 2000 };
+export const RETICLE = { color: '#ff89ba', size: 22 };
+export const DEFAULTS = { survey: 'P/DSS2/color', fov: 1.5 };
